requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-auth>=2.17.0
google-auth-oauthlib>=0.8.0
requests>=2.28.0

